requirements.txt
Flask==2.0.2
gunicorn==20.1.0
requests==2.26.0
python-dotenv==0.19.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
WTForms==3.0.0
email-validator==1.1.3
passlib==1.7.4
PyJWT==2.1.0
stripe==2.65.0
sqlalchemy==1.4.27
psycopg2-binary==2.9.3
openai==0.10.2
eleven==0.0.1
# Note: The version of the eleven library is fictional as of the knowledge cutoff in 2023. Replace with the correct version when available.